refactor: changed Results -> ItunesResponse and ItunesReponse to ArtworkUrl
