+++ app/src/main/java/com/northcoders/recordshop/model/ArtworkUrl.java
@@ -1,0 +1,20 @@
+package com.northcoders.recordshop.model;
+
+public class ArtworkUrl {
+    String artworkUrl100;
+
+    public ArtworkUrl(String artworkUrl100) {
+        this.artworkUrl100 = artworkUrl100;
+    }
+
+    public ArtworkUrl() {
+    }
+
+    public String getArtworkUrl100() {
+        return artworkUrl100;
+    }
+
+    public void setArtworkUrl100(String artworkUrl100) {
+        this.artworkUrl100 = artworkUrl100;
+    }
+}
