+++ app/src/main/java/com/northcoders/recordshop/model/service/ItunesApiService.java
@@ -1,6 +1,6 @@
 package com.northcoders.recordshop.model.service;
 
-import com.northcoders.recordshop.model.Results;
+import com.northcoders.recordshop.model.ItunesResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -9,5 +9,5 @@ import retrofit2.http.Query;
 public interface ItunesApiService {
 
     @GET("search?entity=album&country=gb&limit=1")
-    Call<Results> getAlbumArtworkUrl(@Query("term") String searchQuery);
+    Call<ItunesResponse> getAlbumArtworkUrl(@Query("term") String searchQuery);
 }
