+++ app/src/main/java/com/northcoders/recordshop/ui/mainactivity/MainActivity.java
@@ -14,7 +14,6 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.northcoders.recordshop.R;
 import com.northcoders.recordshop.databinding.ActivityMainBinding;
 import com.northcoders.recordshop.model.Album;
-import com.northcoders.recordshop.model.ItunesResponse;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -90,7 +89,6 @@ public class MainActivity extends AppCompatActivity {
         });
 
     }
-
 
     private void filterList(String titleText, String artistText) {
         filteredAlbumList = new ArrayList<>();
